pyproject.toml
[tool.poetry]
name = "autometabuilder"
version = "0.1.0"
description = "AutoMetabuilder"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "autometabuilder", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
openai = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autometabuilder = "autometabuilder.main:main"
